pyproject.toml
[project]
name = "mystery-agents"
version = "0.5.0"
description = "AI-powered mystery party game generator using LangGraph and LangChain"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Juanje Ojeda", email = "juanje.ojeda@gmail.com"},
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.7",
    "langgraph>=1.0.3",
    "langchain-core>=1.0.5",
    "langchain-google-genai>=3.0.3",
    "pydantic>=2.12.4",
    "click>=8.3.0",
    "weasyprint>=66.0",
    "markdown>=3.10",
    "pillow>=11.0.0",
    "aiohttp>=3.11.0",
    "pyyaml>=6.0.3",
    "babel>=2.13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
mystery-agents = "mystery_agents.cli:generate"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = "weasyprint.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "markdown.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain_google_genai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langgraph.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_ollama: marks tests that require Ollama to be running",
    "requires_gemini: marks tests that require Google Gemini API key",
]
filterwarnings = [
    "ignore::UserWarning:weasyprint.*",
    "ignore:instantiateVariableFont.*:UserWarning",
]

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
    "types-PyYAML>=6.0.0",
]

